[#10] Remove trailing '. Wikipedia' from speech results.
When a result is sourced from Wikipedia, the card has a trailing ' Wikipedia' which, as is mentioned in #10 ruins the effect.

This simple addition strips any '. Wikipedia' text from the end of a result. If Wikipedia was the last word of a result it would be ' Wikipedia.' which breaks the regex check.

This commit also includes an updated (and initially tested) `Archive.zip`. Although I haven't exhaustively tested each feature in the script, I have confirmed that this does infact remove 'Wikipedia' as expected.

A simple test for this is:
```
Alexa, Ask Google who is Barack Obama?
```

You will be able to see that the trailing 'Wikipedia' is removed with this update.